Fit models to the simulated data

Takes simulated data from generate and applies functions to it, usually model-fitting functions.

fit(
  obj,
  ...,
  quiet = TRUE,
  warn_on_error = TRUE,
  stop_on_error = FALSE,
  debug = FALSE,
  .progress = FALSE,
  .options = furrr_options()
)

Arguments

obj
a simpr_specify object--the simulated data from generate--or an include object

...
purrr-style formula functions used for computing on the simulated data. See Details and Examples.

quiet
Should simulation errors be broadcast to the user as they occur?

warn_on_error
Should there be a warning when simulation errors occur?

stop_on_error
Should the simulation stop immediately when simulation errors occur?

debug
Run simulation in debug mode, allowing objects, etc. to be explored for each generated variable specification.

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans .

.options
The future specific options to use with the workers when using futures. This must be the result from a call to furrr_options().

Value

a simpr_gen object with additional list-columns for the output of the provided functions (e.g. model outputs). Just like the output of generate, there is one row per repetition per combination of metaparameters, and the columns are the repetition number rep, the metaparameter names, the simulated data sim, with additional columns for the function outputs specified in ....

Details

This is the fourth step in the simulation process: after generating the simulation data, apply functions such as fitting a statistical model to the data. The output is often then passed to tidy_fits or glance_fits to extract relevant parameters from the object, based on tidy and glance from the broom package.

Similar to specify, the ... arguments uses purrr-style formula functions to specify fitting models to the data. The functions are computed within each simulation cell, so dataset names are often unnecessary: for instance, to compute regressions on each cell, you could specify fit(linear_model = ~lm(y ~ x + z). If your modeling function requires a reference to the full dataset, use ., e.g. fit(linear_model = ~lm(y ~ x + z, data = .). These equivalent specifications would compute linear models on each simulation cell if there are variables x, y, and z specified in specify.

Examples

## Generate data to fit models
simple_linear_data = specify(x1 = ~ 2 + rnorm(n),
                               y = ~ 5 + 3*x1 + rnorm(n, 0, sd = 0.5)) %>%
  define(n = 100:101) %>%
  generate(2)

## Fit with a single linear term
linear_fit = simple_linear_data %>%
  fit(linear = ~lm(y ~ x1, data = .))

linear_fit # first fit element also prints

#> full tibble
#> --------------------------
#> # A tibble: 4 × 5
#>   .sim_id     n   rep sim                linear
#>     <int> <int> <int> <list>             <list>
#> 1       1   100     1 <tibble [100 × 2]> <lm>  
#> 2       2   101     1 <tibble [101 × 2]> <lm>  
#> 3       3   100     2 <tibble [100 × 2]> <lm>  
#> 4       4   101     2 <tibble [101 × 2]> <lm>  
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>       x1     y
#>    <dbl> <dbl>
#>  1 1.88   9.89
#>  2 2.86  14.1 
#>  3 1.57   9.88
#>  4 3.38  15.1 
#>  5 2.26  11.6 
#>  6 1.29   8.71
#>  7 0.729  7.13
#>  8 3.04  14.4 
#>  9 1.43  10.4 
#> 10 1.47  10.7 
#> # … with 90 more rows
#> 
#> linear[[1]]
#> --------------------------
#> 
#> Call:
#> lm(formula = y ~ x1, data = .)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>       4.981        3.032  
#> 
#> 

## Each element of $linear is a model object
linear_fit$linear

#> [[1]]
#> 
#> Call:
#> lm(formula = y ~ x1, data = .)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>       4.981        3.032  
#> 
#> 
#> [[2]]
#> 
#> Call:
#> lm(formula = y ~ x1, data = .)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>       5.198        2.929  
#> 
#> 
#> [[3]]
#> 
#> Call:
#> lm(formula = y ~ x1, data = .)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>       4.922        3.055  
#> 
#> 
#> [[4]]
#> 
#> Call:
#> lm(formula = y ~ x1, data = .)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>       5.122        2.940  
#> 
#> 

## We can fit multiple models to the same data
multi_fit = simple_linear_data %>%
  fit(linear = ~lm(y ~ x1, data = .),
      quadratic = ~lm(y ~ x1 + I(x1^2), data = .))

## Two columns, one for each model
multi_fit

#> full tibble
#> --------------------------
#> # A tibble: 4 × 6
#>   .sim_id     n   rep sim                linear quadratic
#>     <int> <int> <int> <list>             <list> <list>   
#> 1       1   100     1 <tibble [100 × 2]> <lm>   <lm>     
#> 2       2   101     1 <tibble [101 × 2]> <lm>   <lm>     
#> 3       3   100     2 <tibble [100 × 2]> <lm>   <lm>     
#> 4       4   101     2 <tibble [101 × 2]> <lm>   <lm>     
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>       x1     y
#>    <dbl> <dbl>
#>  1 1.88   9.89
#>  2 2.86  14.1 
#>  3 1.57   9.88
#>  4 3.38  15.1 
#>  5 2.26  11.6 
#>  6 1.29   8.71
#>  7 0.729  7.13
#>  8 3.04  14.4 
#>  9 1.43  10.4 
#> 10 1.47  10.7 
#> # … with 90 more rows
#> 
#> linear[[1]]
#> --------------------------
#> 
#> Call:
#> lm(formula = y ~ x1, data = .)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>       4.981        3.032  
#> 
#> 
#> quadratic[[1]]
#> --------------------------
#> 
#> Call:
#> lm(formula = y ~ x1 + I(x1^2), data = .)
#> 
#> Coefficients:
#> (Intercept)           x1      I(x1^2)  
#>      4.6417       3.4438      -0.1019  
#> 
#> 

## Again, each element is a model object
multi_fit$quadratic

#> [[1]]
#> 
#> Call:
#> lm(formula = y ~ x1 + I(x1^2), data = .)
#> 
#> Coefficients:
#> (Intercept)           x1      I(x1^2)  
#>      4.6417       3.4438      -0.1019  
#> 
#> 
#> [[2]]
#> 
#> Call:
#> lm(formula = y ~ x1 + I(x1^2), data = .)
#> 
#> Coefficients:
#> (Intercept)           x1      I(x1^2)  
#>     5.17210      2.99702     -0.02076  
#> 
#> 
#> [[3]]
#> 
#> Call:
#> lm(formula = y ~ x1 + I(x1^2), data = .)
#> 
#> Coefficients:
#> (Intercept)           x1      I(x1^2)  
#>     4.74306      3.30813     -0.06536  
#> 
#> 
#> [[4]]
#> 
#> Call:
#> lm(formula = y ~ x1 + I(x1^2), data = .)
#> 
#> Coefficients:
#> (Intercept)           x1      I(x1^2)  
#>     5.23040      2.80192      0.03401  
#> 
#> 

## Can view terms more nicely with tidy_fits
multi_fit %>%
  tidy_fits

#> # A tibble: 20 × 9
#>    .sim_id     n   rep Source    term        estimate std.error statistic  p.value
#>      <int> <int> <int> <chr>     <chr>          <dbl>     <dbl>     <dbl>    <dbl>
#>  1       1   100     1 linear    (Intercept)   4.98      0.135     37.0   2.05e-59
#>  2       1   100     1 linear    x1            3.03      0.0613    49.5   3.93e-71
#>  3       1   100     1 quadratic (Intercept)   4.64      0.199     23.4   1.23e-41
#>  4       1   100     1 quadratic x1            3.44      0.190     18.1   6.50e-33
#>  5       1   100     1 quadratic I(x1^2)      -0.102     0.0446    -2.28  2.46e- 2
#>  6       2   101     1 linear    (Intercept)   5.20      0.0957    54.3   1.66e-75
#>  7       2   101     1 linear    x1            2.93      0.0423    69.2   1.12e-85
#>  8       2   101     1 quadratic (Intercept)   5.17      0.101     51.2   1.46e-72
#>  9       2   101     1 quadratic x1            3.00      0.0928    32.3   4.92e-54
#> 10       2   101     1 quadratic I(x1^2)      -0.0208    0.0250    -0.829 4.09e- 1
#> 11       3   100     2 linear    (Intercept)   4.92      0.110     44.7   5.28e-67
#> 12       3   100     2 linear    x1            3.06      0.0520    58.8   3.04e-78
#> 13       3   100     2 quadratic (Intercept)   4.74      0.148     32.0   2.58e-53
#> 14       3   100     2 quadratic x1            3.31      0.151     21.9   2.62e-39
#> 15       3   100     2 quadratic I(x1^2)      -0.0654    0.0367    -1.78  7.84e- 2
#> 16       4   101     2 linear    (Intercept)   5.12      0.118     43.4   2.93e-66
#> 17       4   101     2 linear    x1            2.94      0.0515    57.1   1.51e-77
#> 18       4   101     2 quadratic (Intercept)   5.23      0.178     29.4   1.78e-50
#> 19       4   101     2 quadratic x1            2.80      0.177     15.9   7.96e-29
#> 20       4   101     2 quadratic I(x1^2)       0.0340    0.0415     0.819 4.15e- 1

## Can view model summaries with glance_fits
multi_fit %>%
  glance_fits

#> # A tibble: 8 × 16
#>   .sim_id     n   rep Source    r.squared adj.r.squared sigma statistic  p.value
#>     <int> <int> <int> <chr>         <dbl>         <dbl> <dbl>     <dbl>    <dbl>
#> 1       1   100     1 linear        0.962         0.961 0.531     2448. 3.93e-71
#> 2       1   100     1 quadratic     0.963         0.963 0.520     1279. 1.93e-70
#> 3       2   101     1 linear        0.980         0.980 0.498     4795. 1.12e-85
#> 4       2   101     1 quadratic     0.980         0.980 0.498     2391. 6.95e-84
#> 5       3   100     2 linear        0.972         0.972 0.518     3458. 3.04e-78
#> 6       3   100     2 quadratic     0.973         0.973 0.512     1769. 4.75e-77
#> 7       4   101     2 linear        0.970         0.970 0.502     3255. 1.51e-77
#> 8       4   101     2 quadratic     0.971         0.970 0.503     1622. 7.73e-76
#> # … with 7 more variables: df <dbl>, logLik <dbl>, AIC <dbl>, BIC <dbl>,
#> #   deviance <dbl>, df.residual <int>, nobs <int>

## Fit functions do not actually need to be any particular kind of model, they
## can be any arbitrary function. However, not all functions will lead to useful
## output with tidy_fits and glance_fits.
add_five_data = simple_linear_data %>%
  fit(add_five = ~ . + 5)

add_five_data

#> full tibble
#> --------------------------
#> # A tibble: 4 × 5
#>   .sim_id     n   rep sim                add_five      
#>     <int> <int> <int> <list>             <list>        
#> 1       1   100     1 <tibble [100 × 2]> <df [100 × 2]>
#> 2       2   101     1 <tibble [101 × 2]> <df [101 × 2]>
#> 3       3   100     2 <tibble [100 × 2]> <df [100 × 2]>
#> 4       4   101     2 <tibble [101 × 2]> <df [101 × 2]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>       x1     y
#>    <dbl> <dbl>
#>  1 1.88   9.89
#>  2 2.86  14.1 
#>  3 1.57   9.88
#>  4 3.38  15.1 
#>  5 2.26  11.6 
#>  6 1.29   8.71
#>  7 0.729  7.13
#>  8 3.04  14.4 
#>  9 1.43  10.4 
#> 10 1.47  10.7 
#> # … with 90 more rows
#> 
#> add_five[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>       x1     y
#>    <dbl> <dbl>
#>  1  6.88  14.9
#>  2  7.86  19.1
#>  3  6.57  14.9
#>  4  8.38  20.1
#>  5  7.26  16.6
#>  6  6.29  13.7
#>  7  5.73  12.1
#>  8  8.04  19.4
#>  9  6.43  15.4
#> 10  6.47  15.7
#> # … with 90 more rows
#>