Fit models to the simulated data

Takes simulated data from generate and applies functions to it, usually model-fitting functions.

# S3 method for simpr_tibble
fit(
  object,
  ...,
  .quiet = TRUE,
  .warn_on_error = TRUE,
  .stop_on_error = FALSE,
  .debug = FALSE,
  .progress = FALSE,
  .options = furrr_options()
)

# S3 method for simpr_spec
fit(
  object,
  ...,
  .quiet = TRUE,
  .warn_on_error = TRUE,
  .stop_on_error = FALSE,
  .debug = FALSE,
  .progress = FALSE,
  .options = furrr_options()
)

Arguments

object

a simpr_tibble object--the simulated data from generate--or an simpr_spec object not yet generated.

...

purrr-style lambda functions used for computing on the simulated data. See Details and Examples.

.quiet

Should simulation errors be broadcast to the user as they occur?

.warn_on_error

Should there be a warning when simulation errors occur? See vignette("Managing simulation errors").

.stop_on_error

Should the simulation stop immediately when simulation errors occur?

.debug

Run simulation in debug mode, allowing objects, etc. to be explored for each attempt to fit objects.

.progress

A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans .

.options

The future specific options to use with the workers when using futures. This must be the result from a call to furrr_options().

Value

a simpr_tibble object with additional list-columns for the output of the provided functions (e.g. model outputs). Just like the output of generate, there is one row per repetition per combination of metaparameters, and the columns are the repetition number rep, the metaparameter names, the simulated data sim, with additional columns for the function outputs specified in .... If per_sim was called previously, fit returns the object to default simpr_tibble mode.

Details

This is the fourth step in the simulation process: after generating the simulation data, apply functions such as fitting a statistical model to the data. The output is often then passed to tidy_fits or glance_fits to extract relevant model estimates from the object.

Similar to specify, the model-fitting ... arguments can be arbitrary R expressions (purrr-style lambda functions, see as_mapper) to specify fitting models to the data. The functions are computed within each simulation cell, so dataset names are generally unnecessary: e.g., to compute regressions on each cell, fit(linear_model = ~ lm(c ~ a + b). If your modeling function requires a reference to the full dataset, use ., e.g. fit(linear_model = ~lm(c ~ a + b, data = .).

Examples

## Generate data to fit models
simple_linear_data = specify(a = ~ 2 + rnorm(n),
                               b = ~ 5 + 3*a + rnorm(n, 0, sd = 0.5)) %>%
  define(n = 100:101) %>%
  generate(2)

## Fit with a single linear term
linear_fit = simple_linear_data %>%
  fit(linear = ~lm(b ~ a, data = .))

linear_fit # first fit element also prints
#> full tibble
#> --------------------------
#> # A tibble: 4 × 5
#>   .sim_id     n   rep sim                linear
#>     <int> <int> <int> <list>             <list>
#> 1       1   100     1 <tibble [100 × 2]> <lm>  
#> 2       2   101     1 <tibble [101 × 2]> <lm>  
#> 3       3   100     2 <tibble [100 × 2]> <lm>  
#> 4       4   101     2 <tibble [101 × 2]> <lm>  
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>        a     b
#>    <dbl> <dbl>
#>  1 1.12   7.74
#>  2 4.62  18.3 
#>  3 2.41  12.7 
#>  4 2.96  14.1 
#>  5 0.711  6.87
#>  6 2.06  10.5 
#>  7 2.08  12.0 
#>  8 2.56  12.7 
#>  9 2.56  12.8 
#> 10 0.180  5.80
#> # … with 90 more rows
#> 
#> linear[[1]]
#> --------------------------
#> 
#> Call:
#> lm(formula = b ~ a, data = .)
#> 
#> Coefficients:
#> (Intercept)            a  
#>       5.091        2.961  
#> 
#> 

## Each element of $linear is a model object
linear_fit$linear
#> [[1]]
#> 
#> Call:
#> lm(formula = b ~ a, data = .)
#> 
#> Coefficients:
#> (Intercept)            a  
#>       5.091        2.961  
#> 
#> 
#> [[2]]
#> 
#> Call:
#> lm(formula = b ~ a, data = .)
#> 
#> Coefficients:
#> (Intercept)            a  
#>       4.994        3.001  
#> 
#> 
#> [[3]]
#> 
#> Call:
#> lm(formula = b ~ a, data = .)
#> 
#> Coefficients:
#> (Intercept)            a  
#>       4.834        3.020  
#> 
#> 
#> [[4]]
#> 
#> Call:
#> lm(formula = b ~ a, data = .)
#> 
#> Coefficients:
#> (Intercept)            a  
#>       5.121        2.958  
#> 
#> 

## We can fit multiple models to the same data
multi_fit = simple_linear_data %>%
  fit(linear = ~lm(b ~ a, data = .),
      quadratic = ~lm(b ~ a + I(a^2), data = .))

## Two columns, one for each model
multi_fit
#> full tibble
#> --------------------------
#> # A tibble: 4 × 6
#>   .sim_id     n   rep sim                linear quadratic
#>     <int> <int> <int> <list>             <list> <list>   
#> 1       1   100     1 <tibble [100 × 2]> <lm>   <lm>     
#> 2       2   101     1 <tibble [101 × 2]> <lm>   <lm>     
#> 3       3   100     2 <tibble [100 × 2]> <lm>   <lm>     
#> 4       4   101     2 <tibble [101 × 2]> <lm>   <lm>     
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>        a     b
#>    <dbl> <dbl>
#>  1 1.12   7.74
#>  2 4.62  18.3 
#>  3 2.41  12.7 
#>  4 2.96  14.1 
#>  5 0.711  6.87
#>  6 2.06  10.5 
#>  7 2.08  12.0 
#>  8 2.56  12.7 
#>  9 2.56  12.8 
#> 10 0.180  5.80
#> # … with 90 more rows
#> 
#> linear[[1]]
#> --------------------------
#> 
#> Call:
#> lm(formula = b ~ a, data = .)
#> 
#> Coefficients:
#> (Intercept)            a  
#>       5.091        2.961  
#> 
#> 
#> quadratic[[1]]
#> --------------------------
#> 
#> Call:
#> lm(formula = b ~ a + I(a^2), data = .)
#> 
#> Coefficients:
#> (Intercept)            a       I(a^2)  
#>    5.112322     2.936397     0.005839  
#> 
#> 

## Again, each element is a model object
multi_fit$quadratic
#> [[1]]
#> 
#> Call:
#> lm(formula = b ~ a + I(a^2), data = .)
#> 
#> Coefficients:
#> (Intercept)            a       I(a^2)  
#>    5.112322     2.936397     0.005839  
#> 
#> 
#> [[2]]
#> 
#> Call:
#> lm(formula = b ~ a + I(a^2), data = .)
#> 
#> Coefficients:
#> (Intercept)            a       I(a^2)  
#>     5.07532      2.88165      0.03077  
#> 
#> 
#> [[3]]
#> 
#> Call:
#> lm(formula = b ~ a + I(a^2), data = .)
#> 
#> Coefficients:
#> (Intercept)            a       I(a^2)  
#>     4.77436      3.10303     -0.02093  
#> 
#> 
#> [[4]]
#> 
#> Call:
#> lm(formula = b ~ a + I(a^2), data = .)
#> 
#> Coefficients:
#> (Intercept)            a       I(a^2)  
#>    5.128320     2.947735     0.002846  
#> 
#> 

## Can view terms more nicely with tidy_fits
multi_fit %>%
  tidy_fits
#> # A tibble: 20 × 9
#>    .sim_id     n   rep Source    term      estimate std.error statistic  p.value
#>      <int> <int> <int> <chr>     <chr>        <dbl>     <dbl>     <dbl>    <dbl>
#>  1       1   100     1 linear    (Interce…  5.09       0.127    40.1    1.33e-62
#>  2       1   100     1 linear    a          2.96       0.0577   51.3    1.29e-72
#>  3       1   100     1 quadratic (Interce…  5.11       0.195    26.2    9.06e-46
#>  4       1   100     1 quadratic a          2.94       0.181    16.3    1.93e-29
#>  5       1   100     1 quadratic I(a^2)     0.00584    0.0407    0.143  8.86e- 1
#>  6       2   101     1 linear    (Interce…  4.99       0.0994   50.3    2.82e-72
#>  7       2   101     1 linear    a          3.00       0.0463   64.9    6.14e-83
#>  8       2   101     1 quadratic (Interce…  5.08       0.133    38.1    1.61e-60
#>  9       2   101     1 quadratic a          2.88       0.138    20.9    8.19e-38
#> 10       2   101     1 quadratic I(a^2)     0.0308     0.0335    0.917  3.61e- 1
#> 11       3   100     2 linear    (Interce…  4.83       0.112    43.1    1.62e-65
#> 12       3   100     2 linear    a          3.02       0.0496   60.8    1.18e-79
#> 13       3   100     2 quadratic (Interce…  4.77       0.153    31.2    2.38e-52
#> 14       3   100     2 quadratic a          3.10       0.152    20.4    6.16e-37
#> 15       3   100     2 quadratic I(a^2)    -0.0209     0.0361   -0.579  5.64e- 1
#> 16       4   101     2 linear    (Interce…  5.12       0.103    49.7    7.48e-72
#> 17       4   101     2 linear    a          2.96       0.0483   61.3    1.48e-80
#> 18       4   101     2 quadratic (Interce…  5.13       0.142    36.0    2.38e-58
#> 19       4   101     2 quadratic a          2.95       0.154    19.2    5.53e-35
#> 20       4   101     2 quadratic I(a^2)     0.00285    0.0391    0.0728 9.42e- 1

## Can view model summaries with glance_fits
multi_fit %>%
  glance_fits
#> # A tibble: 8 × 16
#>   .sim_id     n   rep Source    r.squared adj.r.squared sigma statistic  p.value
#>     <int> <int> <int> <chr>         <dbl>         <dbl> <dbl>     <dbl>    <dbl>
#> 1       1   100     1 linear        0.964         0.964 0.509     2632. 1.29e-72
#> 2       1   100     1 quadratic     0.964         0.963 0.511     1303. 8.25e-71
#> 3       2   101     1 linear        0.977         0.977 0.490     4210. 6.14e-83
#> 4       2   101     1 quadratic     0.977         0.977 0.491     2102. 3.30e-81
#> 5       3   100     2 linear        0.974         0.974 0.511     3702. 1.18e-79
#> 6       3   100     2 quadratic     0.974         0.974 0.513     1839. 7.60e-78
#> 7       4   101     2 linear        0.974         0.974 0.477     3759. 1.48e-80
#> 8       4   101     2 quadratic     0.974         0.974 0.480     1860. 1.14e-78
#> # … with 7 more variables: df <dbl>, logLik <dbl>, AIC <dbl>, BIC <dbl>,
#> #   deviance <dbl>, df.residual <int>, nobs <int>

## Fit functions do not actually need to be any particular kind of model, they
## can be any arbitrary function. However, not all functions will lead to useful
## output with tidy_fits and glance_fits.
add_five_data = simple_linear_data %>%
  fit(add_five = ~ . + 5)  ## adds 5 to every value in dataset

add_five_data
#> full tibble
#> --------------------------
#> # A tibble: 4 × 5
#>   .sim_id     n   rep sim                add_five      
#>     <int> <int> <int> <list>             <list>        
#> 1       1   100     1 <tibble [100 × 2]> <df [100 × 2]>
#> 2       2   101     1 <tibble [101 × 2]> <df [101 × 2]>
#> 3       3   100     2 <tibble [100 × 2]> <df [100 × 2]>
#> 4       4   101     2 <tibble [101 × 2]> <df [101 × 2]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>        a     b
#>    <dbl> <dbl>
#>  1 1.12   7.74
#>  2 4.62  18.3 
#>  3 2.41  12.7 
#>  4 2.96  14.1 
#>  5 0.711  6.87
#>  6 2.06  10.5 
#>  7 2.08  12.0 
#>  8 2.56  12.7 
#>  9 2.56  12.8 
#> 10 0.180  5.80
#> # … with 90 more rows
#> 
#> add_five[[1]]
#> --------------------------
#> # A tibble: 100 × 2
#>        a     b
#>    <dbl> <dbl>
#>  1  6.12  12.7
#>  2  9.62  23.3
#>  3  7.41  17.7
#>  4  7.96  19.1
#>  5  5.71  11.9
#>  6  7.06  15.5
#>  7  7.08  17.0
#>  8  7.56  17.7
#>  9  7.56  17.8
#> 10  5.18  10.8
#> # … with 90 more rows
#>