Generate simulated data from specification

Use specification from specify or define to produce simulated data.

generate(
  obj,
  reps,
  ...,
  sim_name = "sim",
  .quiet = TRUE,
  .warn_on_error = TRUE,
  .stop_on_error = FALSE,
  .debug = FALSE,
  .progress = FALSE,
  .options = furrr_options(seed = TRUE)
)

Arguments

obj
a simpr_spec object generated by define or specify, containing the specifications of the simulation

reps
number of replications to run (a whole number greater than 0)

...
filtering criteria for which rows to simulate, passed to filter. This is useful for reproducing just a few selected rows of a simulation without needing to redo the entire simulation.

sim_name
name of the list-column to be created, containing simulation results. Default is "sim"

.quiet
Should simulation errors be broadcast to the user as they occur?

.warn_on_error
Should there be a warning when simulation errors occur?

.stop_on_error
Should the simulation stop immediately when simulation errors occur?

.debug
Run simulation in debug mode, allowing objects, etc. to be explored for each generated variable specification.

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans.

.options
The future specific options to use with the workers when using futures. This must be the result from a call to furrr_options(seed = TRUE).

Value

a simpr_sims object, which is a tibble with a row for each repetition (a total of rep repetitions) for each combination of metaparameters and some extra metadata used by fit. The columns are rep for the repetition number, the names of the metaparameters, and a list-column (named by the argument sim_name) containing the dataset for each repetition and metaparameter combination. simpr_sims objects can be manipulated elementwise by dplyr and tidyr verbs: the command is applied to each element of the simulation list-column.

Details

This is the third step in the simulation process: after specifying the population model and defining the metaparameters, generate is the workhorse function that actually generates the simulated datasets, one for each replication, for each combination of metaparameters. You likely want to use the output of generate to fit model(s) with fit.

Errors you get using this function usually have to do with how you specified the simulation in specify and define.

See also

Examples

meta_list_out = specify(x = ~ MASS::mvrnorm(n, rep(0, 2), Sigma = S)) %>%
  define(n = c(10, 20, 30),
       S = list(independent = diag(2), correlated = diag(2) + 2)) %>%
  generate(3)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

 ## View overall structure of the result and a single simulation output
 meta_list_out

#> full tibble
#> --------------------------
#> # A tibble: 18 × 6
#>    .sim_id     n S_index       rep S             sim              
#>      <int> <dbl> <chr>       <int> <list>        <list>           
#>  1       1    10 independent     1 <dbl [2 × 2]> <tibble [10 × 2]>
#>  2       2    20 independent     1 <dbl [2 × 2]> <tibble [20 × 2]>
#>  3       3    30 independent     1 <dbl [2 × 2]> <tibble [30 × 2]>
#>  4       4    10 correlated      1 <dbl [2 × 2]> <tibble [10 × 2]>
#>  5       5    20 correlated      1 <dbl [2 × 2]> <tibble [20 × 2]>
#>  6       6    30 correlated      1 <dbl [2 × 2]> <tibble [30 × 2]>
#>  7       7    10 independent     2 <dbl [2 × 2]> <tibble [10 × 2]>
#>  8       8    20 independent     2 <dbl [2 × 2]> <tibble [20 × 2]>
#>  9       9    30 independent     2 <dbl [2 × 2]> <tibble [30 × 2]>
#> 10      10    10 correlated      2 <dbl [2 × 2]> <tibble [10 × 2]>
#> 11      11    20 correlated      2 <dbl [2 × 2]> <tibble [20 × 2]>
#> 12      12    30 correlated      2 <dbl [2 × 2]> <tibble [30 × 2]>
#> 13      13    10 independent     3 <dbl [2 × 2]> <tibble [10 × 2]>
#> 14      14    20 independent     3 <dbl [2 × 2]> <tibble [20 × 2]>
#> 15      15    30 independent     3 <dbl [2 × 2]> <tibble [30 × 2]>
#> 16      16    10 correlated      3 <dbl [2 × 2]> <tibble [10 × 2]>
#> 17      17    20 correlated      3 <dbl [2 × 2]> <tibble [20 × 2]>
#> 18      18    30 correlated      3 <dbl [2 × 2]> <tibble [30 × 2]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 10 × 2
#>         x_1     x_2
#>       <dbl>   <dbl>
#>  1 -0.324    0.458 
#>  2  0.572    1.31  
#>  3  0.622   -1.46  
#>  4  1.63    -1.72  
#>  5 -0.00886  0.673 
#>  6  0.468    1.59  
#>  7 -0.349    0.905 
#>  8 -0.411   -0.612 
#>  9  0.218    0.0715
#> 10 -0.266   -0.536 
#> 

 ## Changing reps will change the number of replications and thus the number of
 ## rows in the output
 meta_list_2 = specify(x = ~ MASS::mvrnorm(n, rep(0, 2), Sigma = S)) %>%
  define(n = c(10, 20, 30),
       S = list(independent = diag(2), correlated = diag(2) + 2)) %>%
  generate(4)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

 meta_list_2

#> full tibble
#> --------------------------
#> # A tibble: 24 × 6
#>    .sim_id     n S_index       rep S             sim              
#>      <int> <dbl> <chr>       <int> <list>        <list>           
#>  1       1    10 independent     1 <dbl [2 × 2]> <tibble [10 × 2]>
#>  2       2    20 independent     1 <dbl [2 × 2]> <tibble [20 × 2]>
#>  3       3    30 independent     1 <dbl [2 × 2]> <tibble [30 × 2]>
#>  4       4    10 correlated      1 <dbl [2 × 2]> <tibble [10 × 2]>
#>  5       5    20 correlated      1 <dbl [2 × 2]> <tibble [20 × 2]>
#>  6       6    30 correlated      1 <dbl [2 × 2]> <tibble [30 × 2]>
#>  7       7    10 independent     2 <dbl [2 × 2]> <tibble [10 × 2]>
#>  8       8    20 independent     2 <dbl [2 × 2]> <tibble [20 × 2]>
#>  9       9    30 independent     2 <dbl [2 × 2]> <tibble [30 × 2]>
#> 10      10    10 correlated      2 <dbl [2 × 2]> <tibble [10 × 2]>
#> # … with 14 more rows
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 10 × 2
#>        x_1     x_2
#>      <dbl>   <dbl>
#>  1 -1.03    0.150 
#>  2  0.904   0.955 
#>  3 -0.702  -0.231 
#>  4 -0.0770 -0.0962
#>  5  1.51    0.215 
#>  6  0.311   0.387 
#>  7  2.34   -0.0533
#>  8 -3.02    0.982 
#>  9 -1.65   -0.267 
#> 10 -0.172  -0.820 
#> 

 ## Fitting, tidying functions can be included in this step by running those functions and then
 ## generate.  This can save computation time when doing large
 ## simulations, especially with parallel processing
 meta_list_generate_after = specify(x = ~ MASS::mvrnorm(n, rep(0, 2), Sigma = S)) %>%
  define(n = c(10, 20, 30),
       S = list(independent = diag(2), correlated = diag(2) + 2)) %>%
  fit(lm = ~ lm(x_2 ~ x_1, data = .)) %>%
  tidy_fits %>%
  generate(4)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

  meta_list_generate_after

#> # A tibble: 48 × 10
#>    .sim_id     n S_index   rep Source term  estimate std.error statistic p.value
#>      <int> <dbl> <chr>   <int> <chr>  <chr>    <dbl>     <dbl>     <dbl>   <dbl>
#>  1       1    10 indepe…     1 lm     (Int… -0.251       0.260   -0.964  3.64e-1
#>  2       1    10 indepe…     1 lm     x_1   -0.268       0.331   -0.810  4.42e-1
#>  3       2    20 indepe…     1 lm     (Int…  0.364       0.172    2.12   4.86e-2
#>  4       2    20 indepe…     1 lm     x_1   -0.407       0.184   -2.21   4.04e-2
#>  5       3    30 indepe…     1 lm     (Int…  0.00881     0.171    0.0514 9.59e-1
#>  6       3    30 indepe…     1 lm     x_1   -0.321       0.172   -1.87   7.24e-2
#>  7       4    10 correl…     1 lm     (Int… -0.577       0.451   -1.28   2.37e-1
#>  8       4    10 correl…     1 lm     x_1    0.332       0.460    0.722  4.91e-1
#>  9       5    20 correl…     1 lm     (Int… -0.0493      0.206   -0.239  8.14e-1
#> 10       5    20 correl…     1 lm     x_1    0.741       0.120    6.17   8.02e-6
#> # … with 38 more rows

  ## This is equivalent, but may be slower / more memory-intensive
  meta_list_generate_before = meta_list_2 %>%
  fit(lm = ~ lm(x_2 ~ x_1, data = .)) %>%
  tidy_fits