Simpr methods for tidyverse verbs

These are simpr-compatible methods for generic dplyr and tidyr verbs. The user is not expected to call these methods directly.

# S3 method for simpr_sims
add_count(
  x,
  ...,
  wt = NULL,
  sort = FALSE,
  name = NULL,
  .drop = lifecycle::deprecated()
)

# S3 method for simpr_spec
add_count(
  x,
  ...,
  wt = NULL,
  sort = FALSE,
  name = NULL,
  .drop = lifecycle::deprecated()
)

# S3 method for simpr_sims
anti_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for simpr_spec
anti_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for simpr_sims
arrange_(.data, ..., .dots = list())

# S3 method for simpr_spec
arrange_(.data, ..., .dots = list())

# S3 method for simpr_sims
arrange(.data, ..., .by_group = FALSE)

# S3 method for simpr_spec
arrange(.data, ..., .by_group = FALSE)

# S3 method for simpr_sims
as.tbl(x, ...)

# S3 method for simpr_spec
as.tbl(x, ...)

# S3 method for simpr_sims
auto_copy(x, y, copy = FALSE, ...)

# S3 method for simpr_spec
auto_copy(x, y, copy = FALSE, ...)

# S3 method for simpr_sims
collect(x, ...)

# S3 method for simpr_spec
collect(x, ...)

# S3 method for simpr_sims
compute(x, ...)

# S3 method for simpr_spec
compute(x, ...)

# S3 method for simpr_sims
count(x, ..., wt = NULL, sort = FALSE, name = NULL)

# S3 method for simpr_spec
count(x, ..., wt = NULL, sort = FALSE, name = NULL)

# S3 method for simpr_sims
distinct_(.data, ..., .dots, .keep_all = FALSE)

# S3 method for simpr_spec
distinct_(.data, ..., .dots, .keep_all = FALSE)

# S3 method for simpr_sims
distinct(.data, ..., .keep_all = FALSE)

# S3 method for simpr_spec
distinct(.data, ..., .keep_all = FALSE)

# S3 method for simpr_sims
do_(.data, ..., .dots = list())

# S3 method for simpr_spec
do_(.data, ..., .dots = list())

# S3 method for simpr_sims
do(.data, ...)

# S3 method for simpr_spec
do(.data, ...)

# S3 method for simpr_sims
dplyr_col_modify(data, cols)

# S3 method for simpr_spec
dplyr_col_modify(data, cols)

# S3 method for simpr_sims
dplyr_reconstruct(data, template)

# S3 method for simpr_spec
dplyr_reconstruct(data, template)

# S3 method for simpr_sims
dplyr_row_slice(data, i, ...)

# S3 method for simpr_spec
dplyr_row_slice(data, i, ...)

# S3 method for simpr_sims
filter_(.data, ..., .dots = list())

# S3 method for simpr_spec
filter_(.data, ..., .dots = list())

# S3 method for simpr_sims
filter(.data, ..., .preserve = FALSE)

# S3 method for simpr_spec
filter(.data, ..., .preserve = FALSE)

# S3 method for simpr_sims
full_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_spec
full_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_sims
group_by_(.data, ..., .dots = list(), add = FALSE)

# S3 method for simpr_spec
group_by_(.data, ..., .dots = list(), add = FALSE)

# S3 method for simpr_sims
group_by(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))

# S3 method for simpr_spec
group_by(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))

# S3 method for simpr_sims
group_data(.data)

# S3 method for simpr_spec
group_data(.data)

# S3 method for simpr_sims
group_indices_(.data, ..., .dots = list())

# S3 method for simpr_spec
group_indices_(.data, ..., .dots = list())

# S3 method for simpr_sims
group_indices(.data, ...)

# S3 method for simpr_spec
group_indices(.data, ...)

# S3 method for simpr_sims
group_keys(.tbl, ...)

# S3 method for simpr_spec
group_keys(.tbl, ...)

# S3 method for simpr_sims
group_map(.data, .f, ..., .keep = FALSE)

# S3 method for simpr_spec
group_map(.data, .f, ..., .keep = FALSE)

# S3 method for simpr_sims
group_modify(.data, .f, ..., .keep = FALSE)

# S3 method for simpr_spec
group_modify(.data, .f, ..., .keep = FALSE)

# S3 method for simpr_sims
group_nest(.tbl, ..., .key = "data", keep = FALSE)

# S3 method for simpr_spec
group_nest(.tbl, ..., .key = "data", keep = FALSE)

# S3 method for simpr_sims
group_size(x)

# S3 method for simpr_spec
group_size(x)

# S3 method for simpr_sims
group_split(.tbl, ..., .keep = TRUE)

# S3 method for simpr_spec
group_split(.tbl, ..., .keep = TRUE)

# S3 method for simpr_sims
group_trim(.tbl, .drop = dplyr::group_by_drop_default(.tbl))

# S3 method for simpr_spec
group_trim(.tbl, .drop = dplyr::group_by_drop_default(.tbl))

# S3 method for simpr_sims
group_vars(x)

# S3 method for simpr_spec
group_vars(x)

# S3 method for simpr_sims
groups(x)

# S3 method for simpr_spec
groups(x)

# S3 method for simpr_sims
inner_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_spec
inner_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_sims
intersect(x, y, ...)

# S3 method for simpr_spec
intersect(x, y, ...)

# S3 method for simpr_sims
left_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_spec
left_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_sims
mutate_(.data, ..., .dots = list())

# S3 method for simpr_spec
mutate_(.data, ..., .dots = list())

# S3 method for simpr_sims
mutate(.data, ...)

# S3 method for simpr_spec
mutate(.data, ...)

# S3 method for simpr_sims
n_groups(x)

# S3 method for simpr_spec
n_groups(x)

# S3 method for simpr_sims
nest_by(.data, ..., .key = "data", .keep = FALSE)

# S3 method for simpr_spec
nest_by(.data, ..., .key = "data", .keep = FALSE)

# S3 method for simpr_sims
nest_join(x, y, by = NULL, copy = FALSE, keep = FALSE, name = NULL, ...)

# S3 method for simpr_spec
nest_join(x, y, by = NULL, copy = FALSE, keep = FALSE, name = NULL, ...)

# S3 method for simpr_sims
pull(.data, var = -1, name = NULL, ...)

# S3 method for simpr_spec
pull(.data, var = -1, name = NULL, ...)

# S3 method for simpr_sims
relocate(.data, ..., .before = NULL, .after = NULL)

# S3 method for simpr_spec
relocate(.data, ..., .before = NULL, .after = NULL)

# S3 method for simpr_sims
rename_(.data, ..., .dots = list())

# S3 method for simpr_spec
rename_(.data, ..., .dots = list())

# S3 method for simpr_sims
rename_with(.data, .fn, .cols = dplyr::everything(), ...)

# S3 method for simpr_spec
rename_with(.data, .fn, .cols = dplyr::everything(), ...)

# S3 method for simpr_sims
rename(.data, ...)

# S3 method for simpr_spec
rename(.data, ...)

# S3 method for simpr_sims
right_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_spec
right_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

# S3 method for simpr_sims
rows_delete(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_spec
rows_delete(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_sims
rows_insert(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_spec
rows_insert(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_sims
rows_patch(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_spec
rows_patch(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_sims
rows_update(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_spec
rows_update(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_sims
rows_upsert(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_spec
rows_upsert(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)

# S3 method for simpr_sims
rowwise(data, ...)

# S3 method for simpr_spec
rowwise(data, ...)

# S3 method for simpr_sims
same_src(x, y)

# S3 method for simpr_spec
same_src(x, y)

# S3 method for simpr_sims
sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for simpr_spec
sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for simpr_sims
sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for simpr_spec
sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

# S3 method for simpr_sims
select_(.data, ..., .dots = list())

# S3 method for simpr_spec
select_(.data, ..., .dots = list())

# S3 method for simpr_sims
select(.data, ...)

# S3 method for simpr_spec
select(.data, ...)

# S3 method for simpr_sims
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for simpr_spec
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for simpr_sims
setdiff(x, y, ...)

# S3 method for simpr_spec
setdiff(x, y, ...)

# S3 method for simpr_sims
setequal(x, y, ...)

# S3 method for simpr_spec
setequal(x, y, ...)

# S3 method for simpr_sims
slice_(.data, ..., .dots = list())

# S3 method for simpr_spec
slice_(.data, ..., .dots = list())

# S3 method for simpr_sims
slice_head(.data, ..., n, prop)

# S3 method for simpr_spec
slice_head(.data, ..., n, prop)

# S3 method for simpr_sims
slice_max(.data, order_by, ..., n, prop, with_ties = TRUE)

# S3 method for simpr_spec
slice_max(.data, order_by, ..., n, prop, with_ties = TRUE)

# S3 method for simpr_sims
slice_min(.data, order_by, ..., n, prop, with_ties = TRUE)

# S3 method for simpr_spec
slice_min(.data, order_by, ..., n, prop, with_ties = TRUE)

# S3 method for simpr_sims
slice_sample(.data, ..., n, prop, weight_by = NULL, replace = FALSE)

# S3 method for simpr_spec
slice_sample(.data, ..., n, prop, weight_by = NULL, replace = FALSE)

# S3 method for simpr_sims
slice_tail(.data, ..., n, prop)

# S3 method for simpr_spec
slice_tail(.data, ..., n, prop)

# S3 method for simpr_sims
slice(.data, ..., .preserve = FALSE)

# S3 method for simpr_spec
slice(.data, ..., .preserve = FALSE)

# S3 method for simpr_sims
summarise_(.data, ..., .dots = list())

# S3 method for simpr_spec
summarise_(.data, ..., .dots = list())

# S3 method for simpr_sims
summarise(.data, ..., .groups = NULL)

# S3 method for simpr_spec
summarise(.data, ..., .groups = NULL)

# S3 method for simpr_sims
tally(x, wt = NULL, sort = FALSE, name = NULL)

# S3 method for simpr_spec
tally(x, wt = NULL, sort = FALSE, name = NULL)

# S3 method for simpr_sims
tbl_vars(x)

# S3 method for simpr_spec
tbl_vars(x)

# S3 method for simpr_sims
transmute_(.data, ..., .dots = list())

# S3 method for simpr_spec
transmute_(.data, ..., .dots = list())

# S3 method for simpr_sims
transmute(.data, ...)

# S3 method for simpr_spec
transmute(.data, ...)

# S3 method for simpr_sims
ungroup(x, ...)

# S3 method for simpr_spec
ungroup(x, ...)

# S3 method for simpr_sims
union_all(x, y, ...)

# S3 method for simpr_spec
union_all(x, y, ...)

# S3 method for simpr_sims
union(x, y, ...)

# S3 method for simpr_spec
union(x, y, ...)

# S3 method for simpr_sims
complete_(data, cols, fill = list(), ...)

# S3 method for simpr_spec
complete_(data, cols, fill = list(), ...)

# S3 method for simpr_sims
complete(data, ..., fill = list())

# S3 method for simpr_spec
complete(data, ..., fill = list())

# S3 method for simpr_sims
drop_na_(data, vars)

# S3 method for simpr_spec
drop_na_(data, vars)

# S3 method for simpr_sims
drop_na(data, ...)

# S3 method for simpr_spec
drop_na(data, ...)

# S3 method for simpr_sims
expand_(data, dots, ...)

# S3 method for simpr_spec
expand_(data, dots, ...)

# S3 method for simpr_sims
expand(data, ..., .name_repair = "check_unique")

# S3 method for simpr_spec
expand(data, ..., .name_repair = "check_unique")

# S3 method for simpr_sims
extract_(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)

# S3 method for simpr_spec
extract_(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)

# S3 method for simpr_sims
extract(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)

# S3 method for simpr_spec
extract(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)

# S3 method for simpr_sims
fill_(data, fill_cols, .direction = c("down", "up"))

# S3 method for simpr_spec
fill_(data, fill_cols, .direction = c("down", "up"))

# S3 method for simpr_sims
fill(data, ..., .direction = c("down", "up", "downup", "updown"))

# S3 method for simpr_spec
fill(data, ..., .direction = c("down", "up", "downup", "updown"))

# S3 method for simpr_sims
gather_(
  data,
  key_col,
  value_col,
  gather_cols,
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)

# S3 method for simpr_spec
gather_(
  data,
  key_col,
  value_col,
  gather_cols,
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)

# S3 method for simpr_sims
gather(
  data,
  key = "key",
  value = "value",
  ...,
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)

# S3 method for simpr_spec
gather(
  data,
  key = "key",
  value = "value",
  ...,
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)

# S3 method for simpr_sims
nest_legacy(data, ..., .key = "data")

# S3 method for simpr_spec
nest_legacy(data, ..., .key = "data")

# S3 method for simpr_sims
nest(.data, ..., .names_sep = NULL, .key = lifecycle::deprecated())

# S3 method for simpr_spec
nest(.data, ..., .names_sep = NULL, .key = lifecycle::deprecated())

# S3 method for simpr_sims
pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  names_ptypes = list(),
  names_transform = list(),
  names_repair = "check_unique",
  values_to = "value",
  values_drop_na = FALSE,
  values_ptypes = list(),
  values_transform = list(),
  ...
)

# S3 method for simpr_spec
pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  names_ptypes = list(),
  names_transform = list(),
  names_repair = "check_unique",
  values_to = "value",
  values_drop_na = FALSE,
  values_ptypes = list(),
  values_transform = list(),
  ...
)

# S3 method for simpr_sims
pivot_wider(
  data,
  id_cols = NULL,
  id_expand = FALSE,
  names_from = NULL,
  names_prefix = "",
  names_sep = "_",
  names_glue = NULL,
  names_sort = FALSE,
  names_vary = "fastest",
  names_expand = FALSE,
  names_repair = "check_unique",
  values_from = NULL,
  values_fill = NULL,
  values_fn = NULL,
  unused_fn = NULL,
  ...
)

# S3 method for simpr_spec
pivot_wider(
  data,
  id_cols = NULL,
  id_expand = FALSE,
  names_from = NULL,
  names_prefix = "",
  names_sep = "_",
  names_glue = NULL,
  names_sort = FALSE,
  names_vary = "fastest",
  names_expand = FALSE,
  names_repair = "check_unique",
  values_from = NULL,
  values_fill = NULL,
  values_fn = NULL,
  unused_fn = NULL,
  ...
)

# S3 method for simpr_sims
replace_na(data, replace, ...)

# S3 method for simpr_spec
replace_na(data, replace, ...)

# S3 method for simpr_sims
separate_(
  data,
  col,
  into,
  sep = "[^[:alnum:]]+",
  remove = TRUE,
  convert = FALSE,
  extra = "warn",
  fill = "warn",
  ...
)

# S3 method for simpr_spec
separate_(
  data,
  col,
  into,
  sep = "[^[:alnum:]]+",
  remove = TRUE,
  convert = FALSE,
  extra = "warn",
  fill = "warn",
  ...
)

# S3 method for simpr_sims
separate_rows_(data, cols, sep = "[^[:alnum:].]+", convert = FALSE)

# S3 method for simpr_spec
separate_rows_(data, cols, sep = "[^[:alnum:].]+", convert = FALSE)

# S3 method for simpr_sims
separate_rows(data, ..., sep = "[^[:alnum:].]+", convert = FALSE)

# S3 method for simpr_spec
separate_rows(data, ..., sep = "[^[:alnum:].]+", convert = FALSE)

# S3 method for simpr_sims
separate(
  data,
  col,
  into,
  sep = "[^[:alnum:]]+",
  remove = TRUE,
  convert = FALSE,
  extra = "warn",
  fill = "warn",
  ...
)

# S3 method for simpr_spec
separate(
  data,
  col,
  into,
  sep = "[^[:alnum:]]+",
  remove = TRUE,
  convert = FALSE,
  extra = "warn",
  fill = "warn",
  ...
)

# S3 method for simpr_sims
spread_(
  data,
  key_col,
  value_col,
  fill = NA,
  convert = FALSE,
  drop = TRUE,
  sep = NULL
)

# S3 method for simpr_spec
spread_(
  data,
  key_col,
  value_col,
  fill = NA,
  convert = FALSE,
  drop = TRUE,
  sep = NULL
)

# S3 method for simpr_sims
spread(data, key, value, fill = NA, convert = FALSE, drop = TRUE, sep = NULL)

# S3 method for simpr_spec
spread(data, key, value, fill = NA, convert = FALSE, drop = TRUE, sep = NULL)

# S3 method for simpr_sims
unite_(data, col, from, sep = "_", remove = TRUE)

# S3 method for simpr_spec
unite_(data, col, from, sep = "_", remove = TRUE)

# S3 method for simpr_sims
unite(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)

# S3 method for simpr_spec
unite(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)

# S3 method for simpr_sims
unnest_legacy(data, ..., .drop = NA, .id = NULL, .sep = NULL, .preserve = NULL)

# S3 method for simpr_spec
unnest_legacy(data, ..., .drop = NA, .id = NULL, .sep = NULL, .preserve = NULL)

# S3 method for simpr_sims
unnest(
  data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique",
  .drop = lifecycle::deprecated(),
  .id = lifecycle::deprecated(),
  .sep = lifecycle::deprecated(),
  .preserve = lifecycle::deprecated()
)

# S3 method for simpr_spec
unnest(
  data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique",
  .drop = lifecycle::deprecated(),
  .id = lifecycle::deprecated(),
  .sep = lifecycle::deprecated(),
  .preserve = lifecycle::deprecated()
)

Arguments

x

See original function documentation

...

See original function documentation

wt

See original function documentation

sort

See original function documentation

name

See original function documentation

.drop

See original function documentation

y

See original function documentation

by

See original function documentation

copy

See original function documentation

.data

See original function documentation

.dots

See original function documentation

.by_group

See original function documentation

.keep_all

See original function documentation

data

See original function documentation

cols

See original function documentation

template

See original function documentation

i

See original function documentation

.preserve

See original function documentation

suffix

See original function documentation

keep

See original function documentation

add

See original function documentation

.add

See original function documentation

.tbl

See original function documentation

.f

See original function documentation

.keep

See original function documentation

.key

See original function documentation

var

See original function documentation

.before

See original function documentation

.after

See original function documentation

.fn

See original function documentation

.cols

See original function documentation

in_place

See original function documentation

tbl

See original function documentation

size

See original function documentation

replace

See original function documentation

weight

See original function documentation

.env

See original function documentation

n

See original function documentation

prop

See original function documentation

order_by

See original function documentation

with_ties

See original function documentation

weight_by

See original function documentation

.groups

See original function documentation

fill

See original function documentation

vars

See original function documentation

dots

See original function documentation

.name_repair

See original function documentation

col

See original function documentation

into

See original function documentation

regex

See original function documentation

remove

See original function documentation

convert

See original function documentation

fill_cols

See original function documentation

.direction

See original function documentation

key_col

See original function documentation

value_col

See original function documentation

gather_cols

See original function documentation

na.rm

See original function documentation

factor_key

See original function documentation

key

See original function documentation

value

See original function documentation

.names_sep

See original function documentation

names_to

See original function documentation

names_prefix

See original function documentation

names_sep

See original function documentation

names_pattern

See original function documentation

names_ptypes

See original function documentation

names_transform

See original function documentation

names_repair

See original function documentation

values_to

See original function documentation

values_drop_na

See original function documentation

values_ptypes

See original function documentation

values_transform

See original function documentation

id_cols

See original function documentation

id_expand

See original function documentation

names_from

See original function documentation

names_glue

See original function documentation

names_sort

See original function documentation

names_vary

See original function documentation

names_expand

See original function documentation

values_from

See original function documentation

values_fill

See original function documentation

values_fn

See original function documentation

unused_fn

See original function documentation

sep

See original function documentation

extra

See original function documentation

drop

See original function documentation

from

See original function documentation

.id

See original function documentation

.sep

See original function documentation

keep_empty

See original function documentation

ptype

See original function documentation

Value

simpr_sims methods return a

simpr_sims object with the given data transformation applied to each simulation.

simpr_spec methods return a simpr_spec

object that stores the given data transformation, to be executed when

generate

is called.

Details

See original function documentation for details of the functions. Two methods have been created for each tidyr and dplyr generic function: one for simpr_spec objects (generated by specify and define) which are simply stored for later evaluation by generate, and one for simpr_sims objects to perform the operation elementwise on each simulation.

To use these special per-simulation versions of these tidyverse verbs as a part of a simpr simulation, first run per_sim on the object. If you do NOT want to compute only on each simulation but want to return to the default behavior of operating on the entire simulation tibble, use whole_tibble.