Define metaparameters to vary in simulation

Takes the output of specify (a simpr_specify object) and defines the metaparameters for simulation.

define(.x = NULL, ..., .suffix = "_index")

Arguments

.x
a simpr_spec object (the output of specify)

...
metaparameters: named arguments containing vectors or unidimensional lists of objects to be used in the simulation.

.suffix
name of suffix to append onto index column for list metaparameters, "_index" by default. See Details.

Value

a simpr_define object to pass onto generate for the simulation. This is also a tibble containing all simulation conditions that can be examined or altered directly (in case a fully crossed design is not desired).

Details

This is the second step in the simulation process, after specifying the simulated data using specify. The output of define is then passed to generate to actually generate the simulation.

Metaparameters are named arguments that are used in the simulation. Usually, a metaparameter is some kind of vector or list, representing something that is to be systematically varied as a part of the simulation design. Any metaparameter would also appear in the formulas of specify, and thus the simulation changes depending on the value of the metaparameter.

When creating the simulation, simulations for all possible combinations of metaparameters are generated, a fully crossed simulation design.

When one of ... is a list, a new column is generated in the output to generate to serve as the index of the list. This new column will be the name of the list argument, with the suffix argument appended onto the end. So if Y = list(a = 1:2, b = letters[2:3]), and suffix = "_index", the default, a column named Y_index would be added to the output of generate with values "a" and "b".

Examples

# Simple example of setting a metaparameter
simple_meta = specify(x = ~ 1 + rnorm(n)) %>%
  define(n = c(5, 10)) %>%
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula.
#> 
#> To hide this message, or to avoid moving this formula first, use a different variable name.

simple_meta # $sim has a 5-row tibble and a 10-row tibble

#> full tibble
#> --------------------------
#> # A tibble: 2 × 4
#>   .sim_id     n   rep sim              
#>     <int> <dbl> <int> <list>           
#> 1       1     5     1 <tibble [5 × 1]> 
#> 2       2    10     1 <tibble [10 × 1]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 5 × 1
#>       x
#>   <dbl>
#> 1  1.04
#> 2  2.12
#> 3  2.25
#> 4  1.75
#> 5  2.21
#> 

multi_meta = specify(x = ~ mu + rnorm(n)) %>%
  define(n = c(5, 10),
       mu = seq(-1, 1, length.out = 3)) %>%
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula.
#> 
#> To hide this message, or to avoid moving this formula first, use a different variable name.

multi_meta # generates simulations for all combos of n and mu

#> full tibble
#> --------------------------
#> # A tibble: 6 × 5
#>   .sim_id     n    mu   rep sim              
#>     <int> <dbl> <dbl> <int> <list>           
#> 1       1     5    -1     1 <tibble [5 × 1]> 
#> 2       2    10    -1     1 <tibble [10 × 1]>
#> 3       3     5     0     1 <tibble [5 × 1]> 
#> 4       4    10     0     1 <tibble [10 × 1]>
#> 5       5     5     1     1 <tibble [5 × 1]> 
#> 6       6    10     1     1 <tibble [10 × 1]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 5 × 1
#>         x
#>     <dbl>
#> 1 -1.56  
#> 2  0.0716
#> 3 -1.26  
#> 4 -0.971 
#> 5  0.187 
#> 

# meta can handle lists which can contain multiple matrices, etc.
meta_list_out = specify(x = ~ MASS::mvrnorm(n, rep(0, 2), Sigma = S)) %>%
  define(n = c(10, 20, 30),
       S = list(independent = diag(2), correlated = diag(2) + 2)) %>%
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula.
#> 
#> To hide this message, or to avoid moving this formula first, use a different variable name.

meta_list_out # generates S_index column

#> full tibble
#> --------------------------
#> # A tibble: 6 × 6
#>   .sim_id     n S_index       rep S             sim              
#>     <int> <dbl> <chr>       <int> <list>        <list>           
#> 1       1    10 independent     1 <dbl [2 × 2]> <tibble [10 × 2]>
#> 2       2    20 independent     1 <dbl [2 × 2]> <tibble [20 × 2]>
#> 3       3    30 independent     1 <dbl [2 × 2]> <tibble [30 × 2]>
#> 4       4    10 correlated      1 <dbl [2 × 2]> <tibble [10 × 2]>
#> 5       5    20 correlated      1 <dbl [2 × 2]> <tibble [20 × 2]>
#> 6       6    30 correlated      1 <dbl [2 × 2]> <tibble [30 × 2]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 10 × 2
#>        x_1    x_2
#>      <dbl>  <dbl>
#>  1 -0.211  -0.916
#>  2 -0.467   1.14 
#>  3  0.162  -0.178
#>  4  0.912  -3.23 
#>  5 -0.0369 -0.294
#>  6 -0.221  -0.580
#>  7 -3.37   -1.48 
#>  8 -1.60   -0.834
#>  9 -0.413   0.106
#> 10  0.998  -1.36 
#>