Specify data-generating mechanisms

Specify the data-generating mechanisms for the simulation using purrr-style lambda functions.

# S3 method for formula
specify(x = NULL, ..., use_names = TRUE, sep = "_")

# S3 method for simpr_spec
specify(x, ..., use_names = TRUE, sep = "_")

Arguments

x
a simpr_spec object (the output of define), or NULL to create a new specification

...
purrr-style formula functions used for generating simulation variables.

use_names
Whether to use names generated by the lambda function (TRUE, the default), or to overwrite them with supplied names.

sep
Specify the separator for auto-generating names. See Column naming.

Value

A simpr_specify object which contains the functions needed to generate the simulation; to be passed to define for defining metaparameters or, if there are no metaparameters, directly to generate for generating the simulation.

Also useful is the fact that one can refer to variables in subsequent arguments. So, one could define another variable y that depends on x very simply, e.g. specify(x = ~ 3 + runif(10), y = ~ 2 * x).

Finally, one can also refer to metaparameters that are to be systematically varied in the simulation study. See define and the examples for more details.

Details

This is always the first command in the simulation process, to specify the actual simulated variables.

The ... arguments use an efficient syntax to specify custom functions needed for generating a simulation, based on the purrr package. When producing one variable, one can provide an expression such as specify(x = ~ 3 + runif(10)) instead of defining a custom function.

The results of these functions are brought together into a single tibble, so all functions should produce the same number of rows. However, a function can certainly produce multiple columns as well.

Column naming

Because functions can produce different numbers of columns, some decisions need to be made regarding the names of those columns. If a provided lambda function produces a single column, the name given to the argument becomes the name of the column. If the function already produces column names, then the output will use these names if use_names = TRUE, the default. Otherwise, simpr uses the argument name as a base and auto-numbers the columns. For instance, if the argument x generates a two-column matrix and sep = "_" (the default) the columns will be named x_1and x_2.

Custom names can also be directly provided by provide a double-sided formulas to provide names for multiple columns produced by the function. Double-sided formulas should not be named, since the left-hand side is already providing names. Multiple names can be provided on the left-hand side are provided using cbind, similar to multivariate specifications elsewhere in R, e.g. specify(cbind(x, y) ~ MASS::mvrnorm(5, c(0, 0), Sigma = diag(2))).

Examples

## specify a variable and generate it in the simulation
single_var = specify(x = ~ 1 + rnorm(5)) %>%
  generate(1) # generate a single repetition of the simulation

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

single_var$sim[[1]] # peek at the simulation

#> # A tibble: 5 × 1
#>       x
#>   <dbl>
#> 1 0.377
#> 2 0.667
#> 3 1.90 
#> 4 0.141
#> 5 0.283

two_var = specify(x = ~ 1 + rnorm(5),
                    y = ~ x + 2) %>%
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

two_var$sim[[1]]

#> # A tibble: 5 × 2
#>       x     y
#>   <dbl> <dbl>
#> 1 1.89   3.89
#> 2 1.19   3.19
#> 3 1.14   3.14
#> 4 2.49   4.49
#> 5 0.555  2.55

## Generates x_01 through x_10
autonumber_var = specify(x = ~ MASS::mvrnorm(5, rep(0, 10), Sigma = diag(10))) %>%
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

autonumber_var$sim[[1]]

#> # A tibble: 5 × 10
#>     x_01   x_02   x_03   x_04   x_05    x_06   x_07   x_08   x_09   x_10
#>    <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1  1.45   0.317 -0.806  0.503  0.439  0.0697  1.29   1.00   1.56   1.82 
#> 2  0.335 -1.09  -0.303 -0.604  0.161  0.169  -2.19  -1.44   0.533 -0.358
#> 3 -0.517  0.762 -0.354  1.08  -0.341 -0.562   1.58   0.180 -0.739  1.07 
#> 4 -0.953  1.34   0.377 -0.671 -1.27   1.52   -1.80  -2.14  -0.376 -1.14 
#> 5  1.50   0.858 -2.35  -0.769  1.27  -1.07   -0.219 -0.169 -0.139  0.862

# alternatively, you could use a two-sided formula for names
multi_name = specify(cbind(x, y, z) ~ MASS::mvrnorm(5, rep(0, 3), Sigma = diag(3))) %>%
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

multi_name$sim[[1]]

#> # A tibble: 5 × 3
#>          x       y      z
#>      <dbl>   <dbl>  <dbl>
#> 1  0.625    0.0598 -0.883
#> 2 -0.0540   0.0786  2.62 
#> 3 -0.00915  0.559   0.409
#> 4 -0.217    0.556   0.963
#> 5  1.25    -1.82   -1.29 

# Simple example of setting a metaparameter
simple_meta = specify(x = ~ 1 + rnorm(n)) %>%
  define(n = c(5, 10)) %>% # without this line you would get an error!
  generate(1)

#> Formula specification for 'x' detected. Assuming 'x' is the first formula. To hide this message, or to avoid moving this formula first, use a different variable name.

simple_meta # has two rows now, one for each value of n

#> full tibble
#> --------------------------
#> # A tibble: 2 × 4
#>   .sim_id     n   rep sim              
#>     <int> <dbl> <int> <list>           
#> 1       1     5     1 <tibble [5 × 1]> 
#> 2       2    10     1 <tibble [10 × 1]>
#> 
#> sim[[1]]
#> --------------------------
#> # A tibble: 5 × 1
#>         x
#>     <dbl>
#> 1 -0.132 
#> 2  1.03  
#> 3  2.14  
#> 4 -0.0146
#> 5  1.28  
#> 

simple_meta$sim[[1]]

#> # A tibble: 5 × 1
#>         x
#>     <dbl>
#> 1 -0.132 
#> 2  1.03  
#> 3  2.14  
#> 4 -0.0146
#> 5  1.28  

simple_meta$sim[[2]]

#> # A tibble: 10 × 1
#>        x
#>    <dbl>
#>  1 1.25 
#>  2 0.759
#>  3 0.340
#>  4 0.249
#>  5 2.53 
#>  6 0.503
#>  7 3.55 
#>  8 1.16 
#>  9 2.04 
#> 10 0.578